rename: GlobalResponseDto -> GlobalResponse
