+++ src/main/java/com/example/techit7/article/service/ArticleService.java
@@ -2,7 +2,7 @@ package com.example.techit7.article.service;
 
 import com.example.techit7.article.dto.ArticleRequestDto;
 import com.example.techit7.article.dto.ArticleResponseDto;
-import com.example.techit7.global.dto.GlobalResponseDto;
+import com.example.techit7.global.response.GlobalResponseDto;
 import com.example.techit7.user.entity.SiteUser;
 import org.springframework.data.domain.Page;
