+++ src/main/java/com/example/techit7/article/service/ImageService.java
@@ -6,7 +6,7 @@ import com.example.techit7.article.entity.Image;
 import com.example.techit7.article.errormessage.ErrorMessage;
 import com.example.techit7.article.repository.ArticleRepository;
 import com.example.techit7.article.repository.ImageRepository;
-import com.example.techit7.global.dto.GlobalResponseDto;
+import com.example.techit7.global.response.GlobalResponseDto;
 import com.example.techit7.global.util.FileStore;
 import jakarta.persistence.EntityNotFoundException;
 import java.io.IOException;
