+++ src/main/java/com/example/techit7/user/dto/UserCreateResponseDto.java
@@ -1,5 +1,5 @@
 package com.example.techit7.user.dto;
-import com.example.techit7.global.dto.GlobalResponseDto;
+import com.example.techit7.global.response.GlobalResponseDto;
 
 public class UserCreateResponseDto extends GlobalResponseDto {
